Added better/more sil_global parsing and testing.
Also added `object` and `value_to_bridge_object` parsing, and fixed int parsing (use BigInt for literals)

+++ ca.ualberta.maple.swan.ir/src/scala/ca/ualberta/maple/swan/ir/SWANIR.scala
@@ -89,7 +89,7 @@ object BinaryOperation {
 sealed trait Literal
 object Literal {
   case class string(val value: String) extends Literal
-  case class int(val value: Int) extends Literal
+  case class int(val value: BigInt) extends Literal
   case class float(val value: Float) extends Literal
 }
